Add files via upload
This file is actual UI of Currency Converter with HTML5 and css as a Front End and I have not use the API link as you mentioned in Gmail.
Sir I have created a conversion first from 1. USD to INR  2. Pound to INR 3. Euro to INR 4. Yen to INR 5. INR to INR and 6. SGD to INR this is done by using dropdownlist and also I have created a another conversion from INR to SGD currency conversion.